+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Pixelation shader for retro effect
+- (Optional) pixelation shader for retro effect
+- (Optional) chromatic aberration shader
 
 ## [0.0.10] - 2023-10-04